+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/upstream/CmcdDataTest.java
@@ -36,23 +36,48 @@ import org.junit.runner.RunWith;
 public class CmcdDataTest {
 
   @Test
-  public void createInstance_populatesCmcdHttRequestHeaders() {
+  public void createInstance_withInvalidFactoryState_throwsIllegalStateException() {
+    CmcdConfiguration cmcdConfiguration =
+        CmcdConfiguration.Factory.DEFAULT.createCmcdConfiguration(MediaItem.EMPTY);
+    ExoTrackSelection trackSelection = mock(ExoTrackSelection.class);
+
+    assertThrows(
+        "Track selection must be set",
+        IllegalStateException.class,
+        () ->
+            new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+                .setObjectType(CmcdData.Factory.OBJECT_TYPE_INIT_SEGMENT)
+                .createCmcdData());
+
+    assertThrows(
+        "Buffered duration must be set",
+        IllegalStateException.class,
+        () ->
+            new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+                .setObjectType(CmcdData.Factory.OBJECT_TYPE_AUDIO_ONLY)
+                .setTrackSelection(trackSelection)
+                .setChunkDurationUs(100_000)
+                .createCmcdData());
+
+    assertThrows(
+        "Chunk duration must be set",
+        IllegalStateException.class,
+        () ->
+            new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+                .setObjectType(CmcdData.Factory.OBJECT_TYPE_AUDIO_ONLY)
+                .setTrackSelection(trackSelection)
+                .setBufferedDurationUs(100_000)
+                .createCmcdData());
+  }
+
+  @Test
+  public void createInstance_audioObjectType_setsCorrectHttpHeaders() {
     CmcdConfiguration.Factory cmcdConfigurationFactory =
         mediaItem ->
             new CmcdConfiguration(
                 "sessionId",
                 mediaItem.mediaId,
                 new CmcdConfiguration.RequestConfig() {
-                  @Override
-                  public ImmutableListMultimap<@CmcdConfiguration.HeaderKey String, String>
-                      getCustomData() {
-                    return new ImmutableListMultimap.Builder<String, String>()
-                        .putAll("CMCD-Object", "key-1=1", "key-2-separated-by-multiple-hyphens=2")
-                        .put("CMCD-Request", "key-3=\"stringValue1,stringValue2\"")
-                        .put("CMCD-Status", "key-4=\"stringValue3=stringValue4\"")
-                        .build();
-                  }
-
                   @Override
                   public int getRequestedMaximumThroughputKbps(int throughputKbps) {
                     return 2 * throughputKbps;
@@ -69,15 +94,14 @@ public class CmcdDataTest {
     when(trackSelection.getLatestBitrateEstimate()).thenReturn(500_000L);
     DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
     CmcdData cmcdData =
-        new CmcdData.Factory(
-                cmcdConfiguration,
-                trackSelection,
-                /* bufferedDurationUs= */ 1_760_000,
-                /* playbackRate= */ 2.0f,
-                /* streamingFormat= */ CmcdData.Factory.STREAMING_FORMAT_DASH,
-                /* isLive= */ true,
-                /* didRebuffer= */ true,
-                /* isBufferEmpty= */ false)
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setTrackSelection(trackSelection)
+            .setObjectType(CmcdData.Factory.OBJECT_TYPE_AUDIO_ONLY)
+            .setBufferedDurationUs(1_760_000)
+            .setPlaybackRate(2.0f)
+            .setIsLive(true)
+            .setDidRebuffer(true)
+            .setIsBufferEmpty(false)
             .setChunkDurationUs(3_000_000)
             .createCmcdData();
 
@@ -86,32 +110,23 @@ public class CmcdDataTest {
     assertThat(dataSpec.httpRequestHeaders)
         .containsExactly(
             "CMCD-Object",
-            "br=840,d=3000,key-1=1,key-2-separated-by-multiple-hyphens=2,tb=1000",
+            "br=840,d=3000,ot=a,tb=1000",
             "CMCD-Request",
-            "bl=1800,dl=900,key-3=\"stringValue1,stringValue2\",mtp=500,su",
+            "bl=1800,dl=900,mtp=500,su",
             "CMCD-Session",
             "cid=\"mediaId\",pr=2.00,sf=d,sid=\"sessionId\",st=l",
             "CMCD-Status",
-            "bs,key-4=\"stringValue3=stringValue4\",rtp=1700");
+            "bs,rtp=1700");
   }
 
   @Test
-  public void createInstance_populatesCmcdHttpQueryParameters() {
+  public void createInstance_audioObjectType_setsCorrectQueryParameters() {
     CmcdConfiguration.Factory cmcdConfigurationFactory =
         mediaItem ->
             new CmcdConfiguration(
                 "sessionId",
                 mediaItem.mediaId,
                 new CmcdConfiguration.RequestConfig() {
-                  @Override
-                  public ImmutableListMultimap<@CmcdConfiguration.HeaderKey String, String>
-                      getCustomData() {
-                    return new ImmutableListMultimap.Builder<String, String>()
-                        .put("CMCD-Object", "key-1=1")
-                        .put("CMCD-Request", "key-2=\"stringVälue1,stringVälue2\"")
-                        .build();
-                  }
-
                   @Override
                   public int getRequestedMaximumThroughputKbps(int throughputKbps) {
                     return 2 * throughputKbps;
@@ -129,32 +144,148 @@ public class CmcdDataTest {
     when(trackSelection.getLatestBitrateEstimate()).thenReturn(500_000L);
     DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
     CmcdData cmcdData =
-        new CmcdData.Factory(
-                cmcdConfiguration,
-                trackSelection,
-                /* bufferedDurationUs= */ 1_760_000,
-                /* playbackRate= */ 2.0f,
-                /* streamingFormat= */ CmcdData.Factory.STREAMING_FORMAT_DASH,
-                /* isLive= */ true,
-                /* didRebuffer= */ true,
-                /* isBufferEmpty= */ false)
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setObjectType(CmcdData.Factory.OBJECT_TYPE_AUDIO_ONLY)
+            .setTrackSelection(trackSelection)
+            .setBufferedDurationUs(1_760_000)
+            .setPlaybackRate(2.0f)
+            .setIsLive(true)
+            .setDidRebuffer(true)
+            .setIsBufferEmpty(false)
             .setChunkDurationUs(3_000_000)
             .createCmcdData();
 
     dataSpec = cmcdData.addToDataSpec(dataSpec);
 
-    // Confirm that the values above are URL-encoded
-    assertThat(dataSpec.uri.toString()).doesNotContain("ä");
-    assertThat(dataSpec.uri.toString()).contains(Uri.encode("ä"));
     assertThat(dataSpec.uri.getQueryParameter(CmcdConfiguration.CMCD_QUERY_PARAMETER_KEY))
         .isEqualTo(
-            "bl=1800,br=840,bs,cid=\"mediaId\",d=3000,dl=900,key-1=1,"
-                + "key-2=\"stringVälue1,stringVälue2\",mtp=500,pr=2.00,rtp=1700,sf=d,"
-                + "sid=\"sessionId\",st=l,su,tb=1000");
+            "bl=1800,br=840,bs,cid=\"mediaId\",d=3000,dl=900,mtp=500,ot=a,pr=2.00,"
+                + "rtp=1700,sf=d,sid=\"sessionId\",st=l,su,tb=1000");
   }
 
   @Test
-  public void createInstance_withInvalidNonHyphenatedCustomKey_throwsIllegalStateException() {
+  public void createInstance_manifestObjectType_setsCorrectHttpHeaders() {
+    CmcdConfiguration.Factory cmcdConfigurationFactory =
+        mediaItem ->
+            new CmcdConfiguration(
+                "sessionId", mediaItem.mediaId, new CmcdConfiguration.RequestConfig() {});
+    MediaItem mediaItem = new MediaItem.Builder().setMediaId("mediaId").build();
+    CmcdConfiguration cmcdConfiguration =
+        cmcdConfigurationFactory.createCmcdConfiguration(mediaItem);
+    DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
+    CmcdData cmcdData =
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setObjectType(CmcdData.Factory.OBJECT_TYPE_MANIFEST)
+            .createCmcdData();
+
+    dataSpec = cmcdData.addToDataSpec(dataSpec);
+
+    assertThat(dataSpec.httpRequestHeaders)
+        .containsExactly(
+            "CMCD-Object", "ot=m", "CMCD-Session", "cid=\"mediaId\",sf=d,sid=\"sessionId\"");
+  }
+
+  @Test
+  public void createInstance_manifestObjectType_setsCorrectQueryParameters() {
+    CmcdConfiguration.Factory cmcdConfigurationFactory =
+        mediaItem ->
+            new CmcdConfiguration(
+                "sessionId",
+                mediaItem.mediaId,
+                new CmcdConfiguration.RequestConfig() {},
+                CmcdConfiguration.MODE_QUERY_PARAMETER);
+    MediaItem mediaItem = new MediaItem.Builder().setMediaId("mediaId").build();
+    CmcdConfiguration cmcdConfiguration =
+        cmcdConfigurationFactory.createCmcdConfiguration(mediaItem);
+    DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
+    CmcdData cmcdData =
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setObjectType(CmcdData.Factory.OBJECT_TYPE_MANIFEST)
+            .createCmcdData();
+
+    dataSpec = cmcdData.addToDataSpec(dataSpec);
+
+    assertThat(dataSpec.uri.getQueryParameter(CmcdConfiguration.CMCD_QUERY_PARAMETER_KEY))
+        .isEqualTo("cid=\"mediaId\",ot=m,sf=d,sid=\"sessionId\"");
+  }
+
+  @Test
+  public void createInstance_unsetObjectType_setsCorrectHttpHeaders() {
+    CmcdConfiguration.Factory cmcdConfigurationFactory =
+        mediaItem ->
+            new CmcdConfiguration(
+                "sessionId", mediaItem.mediaId, new CmcdConfiguration.RequestConfig() {});
+    MediaItem mediaItem = new MediaItem.Builder().setMediaId("mediaId").build();
+    CmcdConfiguration cmcdConfiguration =
+        cmcdConfigurationFactory.createCmcdConfiguration(mediaItem);
+    ExoTrackSelection trackSelection = mock(ExoTrackSelection.class);
+    Format format = new Format.Builder().setPeakBitrate(840_000).build();
+    when(trackSelection.getSelectedFormat()).thenReturn(format);
+    when(trackSelection.getTrackGroup())
+        .thenReturn(new TrackGroup(format, new Format.Builder().setPeakBitrate(1_000_000).build()));
+    when(trackSelection.getLatestBitrateEstimate()).thenReturn(500_000L);
+    DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
+    CmcdData cmcdData =
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setTrackSelection(trackSelection)
+            .setPlaybackRate(2.0f)
+            .setIsLive(true)
+            .setDidRebuffer(true)
+            .setIsBufferEmpty(false)
+            .createCmcdData();
+
+    dataSpec = cmcdData.addToDataSpec(dataSpec);
+
+    assertThat(dataSpec.httpRequestHeaders)
+        .containsExactly(
+            "CMCD-Object",
+            "br=840,tb=1000",
+            "CMCD-Request",
+            "mtp=500,su",
+            "CMCD-Session",
+            "cid=\"mediaId\",pr=2.00,sf=d,sid=\"sessionId\",st=l",
+            "CMCD-Status",
+            "bs");
+  }
+
+  @Test
+  public void createInstance_unsetObjectType_setsCorrectQueryParameters() {
+    CmcdConfiguration.Factory cmcdConfigurationFactory =
+        mediaItem ->
+            new CmcdConfiguration(
+                "sessionId",
+                mediaItem.mediaId,
+                new CmcdConfiguration.RequestConfig() {},
+                CmcdConfiguration.MODE_QUERY_PARAMETER);
+    MediaItem mediaItem = new MediaItem.Builder().setMediaId("mediaId").build();
+    CmcdConfiguration cmcdConfiguration =
+        cmcdConfigurationFactory.createCmcdConfiguration(mediaItem);
+    ExoTrackSelection trackSelection = mock(ExoTrackSelection.class);
+    Format format = new Format.Builder().setPeakBitrate(840_000).build();
+    when(trackSelection.getSelectedFormat()).thenReturn(format);
+    when(trackSelection.getTrackGroup())
+        .thenReturn(new TrackGroup(format, new Format.Builder().setPeakBitrate(1_000_000).build()));
+    when(trackSelection.getLatestBitrateEstimate()).thenReturn(500_000L);
+    DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
+    CmcdData cmcdData =
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setTrackSelection(trackSelection)
+            .setPlaybackRate(2.0f)
+            .setIsLive(true)
+            .setDidRebuffer(true)
+            .setIsBufferEmpty(false)
+            .createCmcdData();
+
+    dataSpec = cmcdData.addToDataSpec(dataSpec);
+
+    assertThat(dataSpec.uri.getQueryParameter(CmcdConfiguration.CMCD_QUERY_PARAMETER_KEY))
+        .isEqualTo(
+            "br=840,bs,cid=\"mediaId\",mtp=500,pr=2.00,"
+                + "sf=d,sid=\"sessionId\",st=l,su,tb=1000");
+  }
+
+  @Test
+  public void createInstance_customData_setsCorrectHttpHeaders() {
     CmcdConfiguration.Factory cmcdConfigurationFactory =
         mediaItem ->
             new CmcdConfiguration(
@@ -164,27 +295,94 @@ public class CmcdDataTest {
                   @Override
                   public ImmutableListMultimap<@CmcdConfiguration.HeaderKey String, String>
                       getCustomData() {
+                    return new ImmutableListMultimap.Builder<String, String>()
+                        .putAll("CMCD-Object", "key-1=1", "key-2-separated-by-multiple-hyphens=2")
+                        .put("CMCD-Request", "key-3=\"stringValue1,stringValue2\"")
+                        .put("CMCD-Session", "key-4=0.5")
+                        .put("CMCD-Status", "key-5=\"stringValue3=stringValue4\"")
+                        .build();
+                  }
+                });
+    CmcdConfiguration cmcdConfiguration =
+        cmcdConfigurationFactory.createCmcdConfiguration(MediaItem.EMPTY);
+    DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
+    CmcdData cmcdData =
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setObjectType(CmcdData.Factory.OBJECT_TYPE_MANIFEST)
+            .createCmcdData();
+
+    dataSpec = cmcdData.addToDataSpec(dataSpec);
+
+    assertThat(dataSpec.httpRequestHeaders)
+        .containsExactly(
+            "CMCD-Object",
+            "key-1=1,key-2-separated-by-multiple-hyphens=2,ot=m",
+            "CMCD-Request",
+            "key-3=\"stringValue1,stringValue2\"",
+            "CMCD-Session",
+            "key-4=0.5,sf=d",
+            "CMCD-Status",
+            "key-5=\"stringValue3=stringValue4\"");
+  }
+
+  @Test
+  public void createInstance_customData_setsCorrectQueryParameters() {
+    CmcdConfiguration.Factory cmcdConfigurationFactory =
+        mediaItem ->
+            new CmcdConfiguration(
+                null,
+                null,
+                new CmcdConfiguration.RequestConfig() {
+                  @Override
+                  public ImmutableListMultimap<@CmcdConfiguration.HeaderKey String, String>
+                      getCustomData() {
+                    return new ImmutableListMultimap.Builder<String, String>()
+                        .put("CMCD-Object", "key-1=1")
+                        .put("CMCD-Request", "key-2=\"stringVälue1,stringVälue2\"")
+                        .build();
+                  }
+                },
+                CmcdConfiguration.MODE_QUERY_PARAMETER);
+    CmcdConfiguration cmcdConfiguration =
+        cmcdConfigurationFactory.createCmcdConfiguration(MediaItem.EMPTY);
+    DataSpec dataSpec = new DataSpec.Builder().setUri(Uri.EMPTY).build();
+    CmcdData cmcdData =
+        new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+            .setObjectType(CmcdData.Factory.OBJECT_TYPE_MANIFEST)
+            .createCmcdData();
+
+    dataSpec = cmcdData.addToDataSpec(dataSpec);
+
+    // Confirm that the values above are URL-encoded
+    assertThat(dataSpec.uri.toString()).doesNotContain("ä");
+    assertThat(dataSpec.uri.toString()).contains(Uri.encode("ä"));
+    assertThat(dataSpec.uri.getQueryParameter(CmcdConfiguration.CMCD_QUERY_PARAMETER_KEY))
+        .isEqualTo("key-1=1,key-2=\"stringVälue1,stringVälue2\",ot=m,sf=d");
+  }
+
+  @Test
+  public void createInstance_invalidCustomDataKey_throwsException() {
+    CmcdConfiguration.Factory cmcdConfigurationFactory =
+        mediaItem ->
+            new CmcdConfiguration(
+                null,
+                null,
+                new CmcdConfiguration.RequestConfig() {
+                  @Override
+                  public ImmutableListMultimap<@CmcdConfiguration.HeaderKey String, String>
+                      getCustomData() {
+                    // Invalid non-hyphenated key1
                     return ImmutableListMultimap.of("CMCD-Object", "key1=1");
                   }
                 });
-    MediaItem mediaItem = new MediaItem.Builder().setMediaId("mediaId").build();
     CmcdConfiguration cmcdConfiguration =
-        cmcdConfigurationFactory.createCmcdConfiguration(mediaItem);
-    ExoTrackSelection trackSelection = mock(ExoTrackSelection.class);
-    when(trackSelection.getSelectedFormat()).thenReturn(new Format.Builder().build());
+        cmcdConfigurationFactory.createCmcdConfiguration(MediaItem.EMPTY);
 
     assertThrows(
         IllegalStateException.class,
         () ->
-            new CmcdData.Factory(
-                    cmcdConfiguration,
-                    trackSelection,
-                    /* bufferedDurationUs= */ 0,
-                    /* playbackRate= */ 1.0f,
-                    /* streamingFormat= */ CmcdData.Factory.STREAMING_FORMAT_DASH,
-                    /* isLive= */ true,
-                    /* didRebuffer= */ true,
-                    /* isBufferEmpty= */ false)
+            new CmcdData.Factory(cmcdConfiguration, CmcdData.Factory.STREAMING_FORMAT_DASH)
+                .setObjectType(CmcdData.Factory.OBJECT_TYPE_MANIFEST)
                 .createCmcdData());
   }
 }